Create working version of `storage_objects_insert/4`
Instead of composing our own multipart struct and passing that to the
underlying `Gax` request, the official implementation of
`storage_objects_insert/4` would work if one line was changed: the API
path. Right now the autogenerated code isn't prepending `/upload` to the
API path and so will always fail. By changing that, binary data can be
passed to the function using the optional `:body` and `:name` keys. This
allows me to reduce how much of my own testing I need to do and rely on
Google's APIs a bit more.