Fix problems with file names/paths
One of the things I hate about Waffle is that they are extremely
confusing on what is a filename versus a file "name" versus a path
versus a bunch of stuff. It's not clear cut and the terminology should
be fixed. Regardless, this should fix the problem with incorrect
filenames and file "names" (as in with extension vs without extension).